requirements.txt
alabaster==0.7.12
alembic==1.4.2
attrs==19.3.0
Babel==2.8.0
bcrypt==3.1.7
blinker==1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
configparser==5.0.0
docutils==0.16
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.3
gunicorn==20.0.4
idna==2.9
imagesize==1.2.0
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
more-itertools==8.3.0
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==1.4.6
packaging==20.4
Pillow==7.1.2
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
pycparser==2.20
Pygments==2.6.1
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
requests==2.23.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.17
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
